Refactor cards into Screen.
