+++ src/components/Errors/index.js
@@ -1,4 +1,4 @@
 /* eslint-disable import/prefer-default-export */
 
 export { default as AppErrorBoundary } from './AppErrorBoundary';
-export { default as CardErrorBoundary } from './CardErrorBoundary';
+export { default as ScreenErrorBoundary } from './ScreenErrorBoundary';
